+++ src/pyodide-py/pyodide/console.py
@@ -1,4 +1,4 @@
-from typing import Optional, Callable, Any
+from typing import Optional, Callable, Any, List
 import code
 import io
 import sys
@@ -37,7 +37,7 @@ except ImportError:
         return _dummy_promise
 
 
-__all__ = ["InteractiveConsole"]
+__all__ = ["InteractiveConsole", "repr_shorten", "displayhook"]
 
 
 class _StdStream(io.TextIOWrapper):
@@ -111,6 +111,10 @@ class InteractiveConsole(code.InteractiveConsole):
     persistent_stream_redirection
         Wether or not the std redirection should be kept between calls to
         `runcode`.
+    completion
+        Should completion be used? This preloads the `jedi` module to
+        later be used in `complete`. The underlying promise is set to
+        `self.preloads_complete`.
     """
 
     def __init__(
@@ -119,6 +123,7 @@ class InteractiveConsole(code.InteractiveConsole):
         stdout_callback: Optional[Callable[[str], None]] = None,
         stderr_callback: Optional[Callable[[str], None]] = None,
         persistent_stream_redirection: bool = False,
+        completion=False,
     ):
         super().__init__(locals)
         self._stdout = None
@@ -129,6 +134,10 @@ class InteractiveConsole(code.InteractiveConsole):
         if persistent_stream_redirection:
             self.redirect_stdstreams()
         self.run_complete = _dummy_promise
+        self.preloads_complete = _dummy_promise
+        self._completion = completion
+        if completion:
+            self._init_completion()
 
     def redirect_stdstreams(self):
         """ Toggle stdout/stderr redirections. """
@@ -240,6 +249,39 @@ class InteractiveConsole(code.InteractiveConsole):
         version = platform.python_version()
         build = f"({', '.join(platform.python_build())})"
         return f"Python {version} {build} on WebAssembly VM\n{cprt}"
+
+    def _init_completion(self):
+        """ Initalise the completion system (loading Jedi package)."""
+        self._completion_ready = False
+
+        this = self
+
+        def set_ready(*args):
+            this._completion_ready = True
+
+        def load_jedi(*args):
+            return _load_packages_from_imports("import jedi")
+
+        self.preloads_complete = self.preloads_complete.then(load_jedi).then(set_ready)
+
+    def complete(self, source: str) -> List[Any]:
+        """Use jedi to complete a source from local namespace.
+
+        If completion has not been activated in constructor or if
+        it is not yet available (due to package load), it returns
+        an empty list.
+
+        Returns
+        -------
+        A list of Jedi's Completion objects, sorted by name.
+        """
+        if not self._completion or not self._completion_ready:
+            return []
+
+        # here jedi should be ready
+        import jedi
+
+        return jedi.Interpreter(source, [self.locals]).complete()  # type: ignore
 
 
 def repr_shorten(